Move files from include if not relevant
